[naming pt 2] rename `column_resizer` file to match other utils / non-cell files
