+++ packages/eui/src/components/datagrid/body/header/column_resizer.styles.ts
